customer_money를 int에서 long으로 교체

+++ semiproject3/src/main/java/com/example/semiproject3/entity/CustomerDto.java
@@ -22,7 +22,7 @@ public class CustomerDto {
 	private Date customerBirth;
 	private String customerEmail;
 	private int customerPoint;
-	private int customerMoney;
+	private long customerMoney;
 	private String customerGrade;
 	private Date customerJoin;
 	private Date customerLogin;
